Fix an issue with collections in the autocomplete call
The endpoint expects a comma-separated list of strings as a parameter of the GET call,
but passing the string list to the http client fails to perform the proper stringification.

This results in the autocomplete endpoint ignoring the 'collection_ids' parameter, and
returning results across all collections.

This fix is not elegant, but is a workaround that performs the string conversion before
calling the client.